+++ HISTORY.rst
@@ -14,11 +14,13 @@ New indicators
 
 New features and enhancements
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+* New xclim.testing.open_dataset method to read data from the remote testdata repo.
 
 Bug fixes
 ~~~~~~~~~
 * Fixed a bug in the attributes of `frost_season_length`.
 * `indices.run_length` methods using dates now respect the array's calendar.
+* Worked around an xarray bug in sdba.QuantileDeltaMapping when multidimensionnal arrays are used with linear or cubic interpolation.
 
 Internal changes
 ~~~~~~~~~~~~~~~~~